Only 6 functions to resize
